ax_enums.mojom 1
Explicitly casting all uses of enums from ax_enums.idl. This is
preparatory work so we can convert these enums to mojo in a follow-up.

Bug: 650275

+++ ash/system/audio/volume_view.cc
@@ -104,7 +104,8 @@ class VolumeButton : public ButtonListenerActionableView {
     const bool is_pressed = CrasAudioHandler::Get()->IsOutputMuted();
     node_data->AddIntAttribute(
         ui::AX_ATTR_CHECKED_STATE,
-        is_pressed ? ui::AX_CHECKED_STATE_TRUE : ui::AX_CHECKED_STATE_FALSE);
+        static_cast<int32_t>(is_pressed ? ui::AX_CHECKED_STATE_TRUE
+                                        : ui::AX_CHECKED_STATE_FALSE));
   }
 
   views::ImageView* image_;
